refactor: improve app package structure
