+++ app/options.go
@@ -1,0 +1,21 @@
+package app
+
+import (
+	"github.com/gaborage/go-bricks/config"
+	"github.com/gaborage/go-bricks/database"
+	"github.com/gaborage/go-bricks/logger"
+	"github.com/gaborage/go-bricks/messaging"
+)
+
+// Options contains optional dependencies for creating an App instance
+type Options struct {
+	Database               database.Interface
+	MessagingClient        messaging.Client
+	SignalHandler          SignalHandler
+	TimeoutProvider        TimeoutProvider
+	Server                 ServerRunner
+	ConfigLoader           func() (*config.Config, error)
+	DatabaseConnector      func(*config.DatabaseConfig, logger.Logger) (database.Interface, error)
+	MessagingClientFactory func(string, logger.Logger) messaging.AMQPClient
+	ResourceSource         TenantStore
+}
